WIP: show compressed and uncompressed stats
In order to compute the impact of specific data on the compressed payload, it's not sufficient to look at the % of the uncompressed snippet taken up by that data, because the size of the compressed payload does not scale linearly with the data. Update the "analyze" service to generate estimates for gzip-payload impact as well, by removing each piece of data and recompressing the whole.

Similarly, for gzip, a pie chart isn't accurate because the "parts" don't really add up to the whole (due to the nonlinearity), so switch to a "radar" chart of the % of baseline payload size reduced by removing each piece of data.

Still WIP, many things are very messy!